Login Page Updates
Modified the UI to look better. Added a MODAL! This is for when the user
hits the registration button. It pulls up a registration form with some
basic validation and a fun message if you mistype your password
confirmation. Also added a sub menu to the user icon after login. It now
has a settings button which does nothing and a logout button which will
log you out. Woo!